IDEN-486 Introduce authenticationOnChainSmt circuit, remove idOwnershipBySignature circuits and use authentication circuts instead
