+++ circuits/authOnChainSmt.circom
@@ -1,0 +1,5 @@
+pragma circom 2.0.0;
+
+include "lib/authenticationOnChainSmt.circom";
+
+component main {public [userID,challenge,userState]} = VerifyAuthenticationOnChainSmt(32,32);
