+++ circuits/lib/authentication.circom
@@ -1,6 +1,7 @@
 pragma circom 2.0.0;
 
-include "idOwnershipBySignature.circom";
+include "utils/claimUtils.circom";
+include "utils/treeUtils.circom";
 
 template VerifyAuthentication(IdOwnershipLevels) {
 
@@ -22,28 +23,28 @@ template VerifyAuthentication(IdOwnershipLevels) {
 	signal input challengeSignatureS;
 	
     signal input userState;
-    // we have no constraints for "id" in this circuit, however we introduce "id" input here
+    // we have no constraints for "userID" in this circuit, however we introduce "userID" input here
     // as it serves as public input which should be the same for prover and verifier
     signal input userID;
 
-    component checkIdOwnership = IdOwnershipBySignature(IdOwnershipLevels);
+    component verifyAuthClaim = VerifyAuthClaimAndSignature(IdOwnershipLevels);
+    for (var i=0; i<8; i++) { verifyAuthClaim.authClaim[i] <== userAuthClaim[i]; }
+	for (var i=0; i<IdOwnershipLevels; i++) { verifyAuthClaim.authClaimMtp[i] <== userAuthClaimMtp[i]; }
+	verifyAuthClaim.claimsTreeRoot <== userClaimsTreeRoot;
+	verifyAuthClaim.revTreeRoot <== userRevTreeRoot;
+	for (var i=0; i<IdOwnershipLevels; i++) { verifyAuthClaim.authClaimNonRevMtp[i] <== userAuthClaimNonRevMtp[i]; }
+	verifyAuthClaim.authClaimNonRevMtpNoAux <== userAuthClaimNonRevMtpNoAux;
+	verifyAuthClaim.authClaimNonRevMtpAuxHv <== userAuthClaimNonRevMtpAuxHv;
+	verifyAuthClaim.authClaimNonRevMtpAuxHi <== userAuthClaimNonRevMtpAuxHi;
 
-	checkIdOwnership.userClaimsTreeRoot <== userClaimsTreeRoot;
-	for (var i=0; i<IdOwnershipLevels; i++) { checkIdOwnership.userAuthClaimMtp[i] <== userAuthClaimMtp[i]; }
-    for (var i=0; i<8; i++) { checkIdOwnership.userAuthClaim[i] <== userAuthClaim[i]; }
+    verifyAuthClaim.challengeSignatureS <== challengeSignatureS;
+    verifyAuthClaim.challengeSignatureR8x <== challengeSignatureR8x;
+    verifyAuthClaim.challengeSignatureR8y <== challengeSignatureR8y;
+    verifyAuthClaim.challenge <== challenge;
 
-	checkIdOwnership.userRevTreeRoot <== userRevTreeRoot;
-	for (var i=0; i<IdOwnershipLevels; i++) { checkIdOwnership.userAuthClaimNonRevMtp[i] <== userAuthClaimNonRevMtp[i]; }
-	checkIdOwnership.userAuthClaimNonRevMtpNoAux <== userAuthClaimNonRevMtpNoAux;
-	checkIdOwnership.userAuthClaimNonRevMtpAuxHv <== userAuthClaimNonRevMtpAuxHv;
-	checkIdOwnership.userAuthClaimNonRevMtpAuxHi <== userAuthClaimNonRevMtpAuxHi;
-
-    checkIdOwnership.userRootsTreeRoot <== userRootsTreeRoot;
-
-    checkIdOwnership.challenge <== challenge;
-    checkIdOwnership.challengeSignatureR8x <== challengeSignatureR8x;
-    checkIdOwnership.challengeSignatureR8y <== challengeSignatureR8y;
-    checkIdOwnership.challengeSignatureS <== challengeSignatureS;
-    
-    checkIdOwnership.userState <== userState;
+    component checkUserState = checkIdenStateMatchesRoots();
+    checkUserState.claimsTreeRoot <== userClaimsTreeRoot;
+    checkUserState.revTreeRoot <== userRevTreeRoot;
+    checkUserState.rootsTreeRoot <== userRootsTreeRoot;
+    checkUserState.expectedState <== userState;
 }
