--- test/circuits/idOwnershipBySignatureOnChainSmtTest.circom
@@ -1,5 +1,0 @@
-pragma circom 2.0.0;
-
-include "../../circuits/lib/idOwnershipBySignatureOnChainSmt.circom";
-
-component main = IdOwnershipBySignatureOnChainSmt(32, 32);
