--- test/circuits/idOwnershipBySignatureTest.circom
@@ -1,5 +1,0 @@
-pragma circom 2.0.0;
-
-include "../../circuits/lib/idOwnershipBySignature.circom";
-
-component main = IdOwnershipBySignature(4);
